deduplicate mat3mul,mat3mulv, mul_mat_vec_2 into common/math.h
