+++ src/common/math.h
@@ -36,7 +36,7 @@
 #define DT_M_PI_F (3.14159265358979324f)
 #define DT_M_PI (3.14159265358979324)
 
-#define DT_M_LN2f (0.6931471824646f)
+#define DT_M_LN2f (0.6931471805599453f)
 
 static inline float clamp_range_f(const float x, const float low, const float high)
 {
@@ -88,4 +88,44 @@ fastlog (float x)
 {
   return 0.69314718f * fastlog2 (x);
 }
+
+// multiply 3x3 matrix with 3x1 vector
+// dest needs to be different from v
+static inline void mat3mulv(float *const __restrict__ dest, const float *const mat, const float *const __restrict__ v)
+{
+  for(int k = 0; k < 3; k++)
+  {
+    float x = 0.0f;
+    for(int i = 0; i < 3; i++)
+      x += mat[3 * k + i] * v[i];
+    dest[k] = x;
+  }
+}
+
+// multiply two 3x3 matrices
+// dest needs to be different from m1 and m2
+static inline void mat3mul(float *const __restrict__ dest, const float *const __restrict__ m1, const float *const __restrict__ m2)
+{
+  for(int k = 0; k < 3; k++)
+  {
+    for(int i = 0; i < 3; i++)
+    {
+      float x = 0.0f;
+      for(int j = 0; j < 3; j++)
+        x += m1[3 * k + j] * m2[3 * j + i];
+      dest[3 * k + i] = x;
+    }
+  }
+}
+
+static inline void mul_mat_vec_2(const float *m, const float *p, float *o)
+{
+  o[0] = p[0] * m[0] + p[1] * m[1];
+  o[1] = p[0] * m[2] + p[1] * m[3];
+}
+
+
+// modelines: These editor modelines have been set for all relevant files by tools/update_modelines.sh
+// vim: shiftwidth=2 expandtab tabstop=2 cindent
+// kate: tab-indents: off; indent-width 2; replace-tabs on; indent-mode cstyle; remove-trailing-spaces modified;
